Refactor code structure and update API name representation
The main action performed in this commit is an extensive refactoring of the project structure; files were moved from 'ui-skeleton' and 'tests' to 'ui' and 'server/tests' directories respectively. Furthermore, the representation of the API name was updated in various user interfaces from 'Multi-Shot AI' to 'MultiShot.AI'. Lastly, the import path was corrected in 'server/main.py' and Python SDK settings were added to the '.idea/misc.xml'.