Add ItemList component and update chat UI
A new ItemList component has been added, allowing to show and hide a list of items. This component was integrated to the layout file. The UI for the chat component in MultiShotChat.svelte has been updated, making changes to text appearance and adding functionality for automatic scroll to bottom when new messages appear. The +page.svelte file was also modified to accommodate these changes.